+++ messages/de/admin.php
@@ -333,6 +333,8 @@ Da <strong>kein E-Mail-Versand</strong> eingerichtet ist, musst du die <strong>P
     'motion_type_templ_motionh'   => 'Titel, Antragstext, Begründung',
     'motion_type_templ_appl'      => 'Standard-Vorlage: Bewerbung',
     'motion_type_templ_applh'     => 'Name, Foto, Alter, Geschlecht, Geburtsort, Selbstvorstellung, Unterschrift',
+    'motion_type_templ_pdfappl'   => 'Standard-Vorlage: PDF-Bewerbung',
+    'motion_type_templ_pdfapplh'  => 'Name, PDF-Upload',
     'motion_type_templ_none'      => '- keine Vorlage -',
     'motion_type_singular'        => 'Titel (Einzahl)',
     'motion_type_singular_pl'     => 'Antrag',