+++ models/motionTypeTemplates/PDFApplication.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace app\components\motionTypeTemplates;
+namespace app\models\motionTypeTemplates;
 
 use app\models\db\Consultation;
 use app\models\db\ConsultationMotionType;
@@ -9,23 +9,23 @@ use app\models\supportTypes\ISupportType;
 use app\models\policies\IPolicy;
 use app\models\sectionTypes\ISectionType;
 
-trait Manifesto
+trait PDFApplication
 {
     /**
      * @param Consultation $consultation
      * @return ConsultationMotionType
      */
-    public static function doCreateManifestoType(Consultation $consultation)
+    public static function doCreateApplicationType(Consultation $consultation)
     {
         $type                               = new ConsultationMotionType();
         $type->consultationId               = $consultation->id;
-        $type->titleSingular                = \Yii::t('structure', 'preset_manifesto_singular');
-        $type->titlePlural                  = \Yii::t('structure', 'preset_manifesto_plural');
-        $type->createTitle                  = \Yii::t('structure', 'preset_manifesto_call');
+        $type->titleSingular                = \Yii::t('structure', 'preset_app_singular');
+        $type->titlePlural                  = \Yii::t('structure', 'preset_app_plural');
+        $type->createTitle                  = \Yii::t('structure', 'preset_app_call');
         $type->position                     = 0;
-        $type->policyMotions                = IPolicy::POLICY_ADMINS;
-        $type->policyAmendments             = IPolicy::POLICY_ALL;
-        $type->policyComments               = IPolicy::POLICY_ALL;
+        $type->policyMotions                = IPolicy::POLICY_ALL;
+        $type->policyAmendments             = IPolicy::POLICY_NOBODY;
+        $type->policyComments               = IPolicy::POLICY_NOBODY;
         $type->policySupportMotions         = IPolicy::POLICY_NOBODY;
         $type->policySupportAmendments      = IPolicy::POLICY_NOBODY;
         $type->initiatorsCanMergeAmendments = ConsultationMotionType::INITIATORS_MERGE_NEVER;
@@ -33,13 +33,12 @@ trait Manifesto
         $type->contactPhone                 = ConsultationMotionType::CONTACT_OPTIONAL;
         $type->contactEmail                 = ConsultationMotionType::CONTACT_REQUIRED;
         $type->supportType                  = ISupportType::ONLY_INITIATOR;
-        $type->texTemplateId                = 1;
-        $type->amendmentMultipleParagraphs  = 1;
+        $type->amendmentMultipleParagraphs  = 0;
         $type->motionLikesDislikes          = 0;
         $type->amendmentLikesDislikes       = 0;
         $type->status                       = ConsultationMotionType::STATUS_VISIBLE;
-        $type->layoutTwoCols                = 0;
-        $type->sidebarCreateButton          = 1;
+        $type->layoutTwoCols                = 1;
+        $type->sidebarCreateButton          = 0;
         $type->save();
 
         return $type;
@@ -48,35 +47,35 @@ trait Manifesto
     /**
      * @param ConsultationMotionType $motionType
      */
-    public static function doCreateManifestoSections(ConsultationMotionType $motionType)
+    public static function doCreateApplicationSections(ConsultationMotionType $motionType)
     {
         $section                = new ConsultationSettingsMotionSection();
         $section->motionTypeId  = $motionType->id;
         $section->type          = ISectionType::TYPE_TITLE;
         $section->position      = 0;
         $section->status        = ConsultationSettingsMotionSection::STATUS_VISIBLE;
-        $section->title         = \Yii::t('structure', 'preset_manifesto_title');
+        $section->title         = \Yii::t('structure', 'preset_app_name');
         $section->required      = 1;
         $section->maxLen        = 0;
         $section->fixedWidth    = 0;
         $section->lineNumbers   = 0;
         $section->hasComments   = 0;
-        $section->hasAmendments = 1;
+        $section->hasAmendments = 0;
         $section->positionRight = 0;
         $section->save();
 
         $section                = new ConsultationSettingsMotionSection();
         $section->motionTypeId  = $motionType->id;
-        $section->type          = ISectionType::TYPE_TEXT_SIMPLE;
+        $section->type          = ISectionType::TYPE_PDF;
         $section->position      = 1;
         $section->status        = ConsultationSettingsMotionSection::STATUS_VISIBLE;
-        $section->title         = \Yii::t('structure', 'preset_manifesto_text');
+        $section->title         = \Yii::t('structure', 'preset_app_pdf');
         $section->required      = 1;
         $section->maxLen        = 0;
-        $section->fixedWidth    = 1;
-        $section->lineNumbers   = 1;
-        $section->hasComments   = 1;
-        $section->hasAmendments = 1;
+        $section->fixedWidth    = 0;
+        $section->lineNumbers   = 0;
+        $section->hasComments   = 0;
+        $section->hasAmendments = 0;
         $section->positionRight = 0;
         $section->save();
     }